Type safety and warning fixes; add code-issues.md
Improved type safety and explicit casting throughout C++ code to resolve build warnings related to implicit conversions and truncation. Updated string formatting and file handling for correctness. Replaced deprecated functions and string comparisons with modern equivalents. Adjusted struct field types for color values and improved loop variable scoping. Added code-issues.md to document remaining build warnings and recommended actions for both C++ and C# files. Enhanced exception handling and code clarity.

+++ Visualizer/vis_milk2/milkdropfs.cpp
@@ -5730,9 +5730,9 @@ void CPlugin::ShowSongTitleAnim(int w, int h, float fProgress, int supertextInde
     minY = centerY - halfHeight * 0.8f * m_supertexts[supertextIndex].fBoxTop;
     maxY = centerY + halfHeight * 0.8f * m_supertexts[supertextIndex].fBoxBottom;
 
-    int boxColR = std::clamp(m_supertexts[supertextIndex].fBoxColR, 0, 255);
-    int boxColG = std::clamp(m_supertexts[supertextIndex].fBoxColG, 0, 255);
-    int boxColB = std::clamp(m_supertexts[supertextIndex].fBoxColB, 0, 255);
+    int boxColR = static_cast<int>(std::clamp(m_supertexts[supertextIndex].fBoxColR, 0.0f, 255.0f));
+    int boxColG = static_cast<int>(std::clamp(m_supertexts[supertextIndex].fBoxColG, 0.0f, 255.0f));
+    int boxColB = static_cast<int>(std::clamp(m_supertexts[supertextIndex].fBoxColB, 0.0f, 255.0f));
 
     D3DCOLOR boxCol = D3DCOLOR_ARGB(boxAlpha, boxColR, boxColG, boxColB);
